Forcibly remove corpus size cap on once_locally
Since once_locally wasn't passing its own value to the size cap, it was
picking up the global tester's size cap, which is generally not what we
want.  It now specifically passes a negative value to disable the cap.

Eventually, we might want to have this configurable as a parameter.